Integrating Data
This is my third attempt to make this site. TrainingHelper repo is my second attempt but because of role implementation I found myself in a dead end. So I decided to start anew and first implement the database and role logic. In this commit I insert my other half of the site's logic, like : Controllers, Views, Models and such.
Later this day I will test if the site works and everything is alright